server/requirements.txt
python-dotenv
yfinance
finvizfinance
Flask[async]
Flask-Cors
requests
beautifulsoup4
gunicorn
gevent
google-cloud-firestore
google-cloud-secret-manager
firebase-admin
google-auth